+++ src/graphql/resolvers/catalog.js
@@ -1,10 +1,10 @@
 export default app => {
   return {
     Query: {
-      statistics: async (obj, args) => {
+      statisticsCatalog: async (obj, args) => {
         return app.service('çatalog/statistics').find({ query: args })
       },
-      measures: async (obj, args) => {
+      measuresCatalog: async (obj, args) => {
         return app.service('catalog/measures').find({ query: args })
       }
     }