+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## Added
-- `statistics` and `measures` GraphQL queries to inspect and query the data schema
+- `statisticsCatalog` and `measuresCatalog` GraphQL queries to inspect and query the data schema
 - The server now downloads `info.json` on startup and imports the latest schema from GenesAPI `schema.json` and `names.json` instead of previously bundled and static files
 - Query and Schema tests